+++ Flora/src/Flora/Scene/Components.h
@@ -10,7 +10,7 @@ namespace Flora {
 	struct TagComponent {
 		std::string Tag;
 		TagComponent() = default;
-		TagComponent(const TagComponent&) = default;
+		TagComponent(const TagComponent& other) { Tag = other.Tag; }
 		TagComponent(const std::string& tag) : Tag(tag) {}
 	};
 
@@ -19,7 +19,11 @@ namespace Flora {
 		glm::vec3 Rotation = { 0.0f, 0.0f, 0.0f };
 		glm::vec3 Scale = { 1.0f, 1.0f, 1.0f };
 		TransformComponent() = default;
-		TransformComponent(const TransformComponent&) = default;
+		TransformComponent(const TransformComponent& other) {
+			Translation = other.Translation;
+			Rotation = other.Rotation;
+			Scale = other.Scale;
+		}
 		TransformComponent(const glm::vec3& translation) : Translation(translation) {}
 		glm::mat4 GetTransform() const {
 			glm::mat4 rotation = glm::toMat4(glm::quat(Rotation));
@@ -50,7 +54,26 @@ namespace Flora {
 		std::string Path = "NULL";
 		std::string Filename = "None Selected";
 		SpriteRendererComponent() = default;
-		SpriteRendererComponent(const SpriteRendererComponent&) = default;
+		SpriteRendererComponent(const SpriteRendererComponent& other) {
+			Color = other.Color;
+			Texture = other.Texture;
+			Type = other.Type;
+			TilingFactor = other.TilingFactor;
+			Rows = other.Rows;
+			Columns = other.Columns;
+			RowCoordinate = other.RowCoordinate;
+			ColumnCoordinate = other.ColumnCoordinate;
+			SubtextureWidth = other.SubtextureWidth;
+			SubtextureHeight = other.SubtextureHeight;
+			Frames = other.Frames;
+			StartFrame = other.StartFrame;
+			EndFrame = other.EndFrame;
+			CurrentFrame = other.CurrentFrame;
+			FPS = other.FPS;
+			FrameCounter = other.FrameCounter;
+			Path = other.Path;
+			Filename = other.Filename;
+		}
 		SpriteRendererComponent(const glm::vec4& color) : Color(color) {}
 	};
 
@@ -59,7 +82,11 @@ namespace Flora {
 		bool Primary = true; // TODO: think about moving to scene
 		bool FixedAspectRatio = false;
 		CameraComponent() = default;
-		CameraComponent(const CameraComponent&) = default;
+		CameraComponent(const CameraComponent& other) {
+			Camera = other.Camera;
+			Primary = other.Primary;
+			FixedAspectRatio = other.FixedAspectRatio;
+		}
 	};
 
 	struct NativeScriptComponent {
@@ -78,10 +105,25 @@ namespace Flora {
 			InstantiateScript = []() { return static_cast<ScriptableEntity*>(new T()); };
 			DestroyScript = [](NativeScriptComponent* nsc) { delete nsc->Instance; nsc->Instance = nullptr; };
 		}
+
+		NativeScriptComponent() = default;
+		NativeScriptComponent(const NativeScriptComponent& other) {
+			Path = other.Path;
+			Filename = other.Filename;
+		}
 	};
 
 	struct ScriptManagerComponent {
 		std::vector<NativeScriptComponent> NativeScripts;
+		ScriptManagerComponent() = default;
+		ScriptManagerComponent(const ScriptManagerComponent& other) {
+			for (int i = 0; i < other.NativeScripts.size(); i++) {
+				NativeScriptComponent nsc;
+				nsc.Path = other.NativeScripts[i].Path;
+				nsc.Filename = other.NativeScripts[i].Filename;
+				NativeScripts.emplace_back(nsc);
+			}
+		}
 	};
 
 	struct ChildComponent {
@@ -93,11 +135,18 @@ namespace Flora {
 				if (Children[i] == entity)
 					Children.erase(Children.begin() + i);
 		}
+		ChildComponent() = default;
 	};
 
 	struct ParentComponent {
 		Entity Parent;
 		bool InheritAll = true;
 		bool InheritTransform = true;
+		ParentComponent() = default;
+		ParentComponent(const ParentComponent& other) {
+			Parent = other.Parent;
+			InheritAll = other.InheritAll;
+			InheritTransform = other.InheritTransform;
+		}
 	};
 }