+++ Flora/src/Flora/Utils/Structures.h
@@ -4,6 +4,23 @@
 #include "Flora/Core/Layer.h"
 
 namespace Flora {
+	enum class Panels {
+		NONE = 0,
+		CONSOLE = 1,
+		CONTENTBROWSER = 2,
+		PROPERTIES = 3,
+		SCENEHIERARCHY = 4,
+		STATS = 5,
+		VIEWPORT = 6
+	};
+
+	struct Clipboard {
+		std::string Filepath = "";
+		Entity Entity = {};
+		bool CutFile = false;
+		bool CutEntity = false;
+	};
+
 	struct EditorParams {
 		Entity SelectedEntity = {};
 		int GizmoType = -1;
@@ -12,5 +29,8 @@ namespace Flora {
 		bool Resized = true;
 		std::vector<std::string> ClosedPanels; //more runtime-efficient way to do this would be with a hashmap
 		float Time; // in seconds
+		Panels FocusedPanel = Panels::NONE;
+		Panels HoveredPanel = Panels::NONE;
+		Clipboard Clipboard;
 	};
 }